Simplify svcctl pkg
