--- exp/pm/internal/svcctl/svcctl_unix.go
@@ -1,8 +1,0 @@
-//go:build linux || darwin
-// +build linux darwin
-
-package svcctl
-
-var (
-	fileExt = ""
-)
