--- exp/pm/internal/svcctl/svcctl_windows.go
@@ -1,5 +1,0 @@
-package svcctl
-
-var (
-	fileExt = ".exe"
-)
